gfx: Add color buffer and depth buffer texture getters
Color buffer is already usable as texture, whereas depth buffer is not. Therefore, the Window instance holds a render-able texture copy of the depth buffer that must be updated manually by the developer by calling Window::updateDepthBufferTexture() when the developer recognizes that a change happened to the depth buffer (clear, draw, etc.) and the texture is needed.

TODO: Cafe: I currently disabled the stencil buffer, I need to get back and find a way to fix that.